CmdParser class updated to take in optional and required parameters separately. the check for required parameters is done internally. corresponding test has been updated.
git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/CBICA_Toolkit/trunk@223 36980a1d-918e-4cb2-ab0a-7661cd27ea28

+++ src/classes/cbicaUtilities.h
@@ -403,7 +403,7 @@ namespace cbica
       std::find(vector_to_search_in.begin(), vector_to_search_in.end(), element_to_search_for);
     if( iterator !=  vector_to_search_in.end() )
     {
-      position = iterator - vector_to_search_in.begin();
+      position = static_cast<int>(iterator - vector_to_search_in.begin());
       return std::make_pair(true, position);
     }
     else
